hakoniwa-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
google-generativeai==0.8.0
anthropic>=0.40.0
python-dotenv==1.0.1
pydantic==2.9.0
openai>=1.0.0
voyageai
numpy
requests
edge-tts
